pyproject.toml
[tool.poetry]
name = "tgbotchatgpt"
version = "0.1.0"
description = ""
authors = ["lap1as <kuzukoleg12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.26.4"
aiogram = "^2.24"
pandas = "^1.5.3"
python-dotenv = "^0.21.1"
langdetect = "^1.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
